+++ packages/plugin-atp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @iqai/plugin-atp
 
+## 0.3.0
+
+### Minor Changes
+
+- b9407e5: Update Eliza packages to 0.25.9
+
 ## 0.2.0
 
 ### Minor Changes